fix: consolidate login into auth service (closes #207) [skip ci]
- Replace auth login page with self-contained file import + paste key UI (dark theme, matches profile LoginForm)
- Delete profile LoginForm.tsx; replace profile /login with redirect to auth /login
- Update all apps (links, coffee, dykil, connections) to use auth/login?next= instead of auth?redirect=
- Profile edit/page/handle pages now point to auth URL for login links
- Add target: ES2020 to auth tsconfig for BigInt support
