+++ apps/profile/app/[handle]/page.tsx
@@ -210,7 +210,7 @@ export default async function ProfilePage({ params }: PageProps) {
             </p>
             {!viewerDid && (
               <a
-                href="/login"
+                href={`${process.env.NEXT_PUBLIC_SERVICE_PREFIX || 'https://'}auth.${process.env.NEXT_PUBLIC_DOMAIN || 'imajin.ai'}/login`}
                 className="inline-block mt-4 px-6 py-2 bg-[#F59E0B] text-black rounded-lg hover:bg-[#D97706] transition font-medium text-sm"
               >
                 Login to see more
